This will run twice, once for x86 and once for x64, I'll make it cache the executable for now as it'll be quicker

+++ release/wix/build_msi.sh
@@ -32,9 +32,12 @@ if [ ! -f "$BINARY" ]; then
   exit 1
 fi
 
-wget -O "$WORKSPACE/ffmpeg.zip" "https://www.gyan.dev/ffmpeg/builds/ffmpeg-release-essentials.zip"
-unzip -d "$WORKSPACE/extracted_ffmpeg" "$WORKSPACE/ffmpeg.zip" "*/ffmpeg.exe"
-cp "$WORKSPACE"/extracted_ffmpeg/ffmpeg-*-essentials_build/bin/ffmpeg.exe "$MSI_OUTPUT_DIR"
+if [ ! -f "$MSI_OUTPUT_DIR"/ffmpeg.exe ]; then
+  wget -O "$WORKSPACE/ffmpeg.zip" "https://www.gyan.dev/ffmpeg/builds/ffmpeg-release-essentials.zip"
+  rm -rf "$WORKSPACE/extracted_ffmpeg"
+  unzip -d "$WORKSPACE/extracted_ffmpeg" "$WORKSPACE/ffmpeg.zip" "*/ffmpeg.exe"
+  cp "$WORKSPACE"/extracted_ffmpeg/ffmpeg-*-essentials_build/bin/ffmpeg.exe "$MSI_OUTPUT_DIR"
+fi
 
 cp "$WORKSPACE"/LICENSE "$WORKSPACE"/README.md "$MSI_OUTPUT_DIR"
 cp "$BINARY" "$MSI_OUTPUT_DIR"
